Fixed a bug for ints temporarily
Bandaid. tested changing my ints to unsigned and we're back to peak performance. What part of this code change did anything? Where were we using ints that caused this?
It almost had to be an indexing into the vector issue. This might be a permanent solution until we surpass the halfway point of our current list.