Added an mpz only implementation
Except for our starting value and some basic counting of equidistant values, we only use mpzs.
This implementation I thought would be slower until we got to near int max or something. It is a little slower MAYBE in the lower thousands but overall WAY WAY faster with the way we are indexing a map for confirmations and reducing a TON of calculations with this ability.