Oh, it wasnt wrong. Added multithreading
We still would have found a magic square of squares in the values checked if there were one with the formula we use.
Added some very small efficiencies to the code.
I should eventually add some output for finding multiple square numbers in the tips of the cross.

Added multiple threads. Unsure why but these calcs seems to NOT instantly max my CPU at 95 degrees. Nice. Got to be the hang on the mutex or something. Still fast.
Need a timer so I can gauge how much faster 1 thread is vs 6 or 12.