Added probably the final check system
I have looked at the data and configurations over and over and determined that the first step is to find equidistant values left of X that at up with 2 values right of X that equal 3X. X isn't directly involved in these first calcs.
Then, we need to find a 4th value that adds with our Left and 1 of the right vals that equals 3X. That would make a magic square of squares, probably.